refactor: change directory tree
