+++ src/config/winston.ts
@@ -1,9 +1,9 @@
-import { createLogger, format, transports } from "winston";
-import "winston-daily-rotate-file";
-import * as fs from "fs";
+import { createLogger, format, transports } from 'winston';
+import 'winston-daily-rotate-file';
+import * as fs from 'fs';
 
-const env = process.env.NODE_ENV || "development";
-const logDir = "logs";
+const env = process.env.NODE_ENV || 'development';
+const logDir = 'logs';
 
 // https://lovemewithoutall.github.io/it/winston-example/
 // Create the log directory if it does not exist
@@ -12,32 +12,34 @@ if (!fs.existsSync(logDir)) {
 }
 
 const dailyRotateFileTransport = new transports.DailyRotateFile({
-  level: "debug",
+  level: 'debug',
   filename: `${logDir}/%DATE%-smart-push.log`,
-  datePattern: "YYYY-MM-DD",
+  datePattern: 'YYYY-MM-DD',
   zippedArchive: true,
-  maxSize: "20m",
-  maxFiles: "14d",
+  maxSize: '20m',
+  maxFiles: '14d',
 });
 
-export const logger = createLogger({
-  level: env === "development" ? "debug" : "info",
+const logger = createLogger({
+  level: env === 'development' ? 'debug' : 'info',
   format: format.combine(
     format.timestamp({
-      format: "YYYY-MM-DD HH:mm:ss",
+      format: 'YYYY-MM-DD HH:mm:ss',
     }),
-    format.json()
+    format.json(),
   ),
   transports: [
     new transports.Console({
-      level: "info",
+      level: 'info',
       format: format.combine(
         format.colorize(),
         format.printf(
-          (info) => `${info.timestamp} ${info.level}: ${info.message}`
-        )
+          (info) => `${info.timestamp} ${info.level}: ${info.message}`,
+        ),
       ),
     }),
     dailyRotateFileTransport,
   ],
 });
+
+export default logger;
